Move settings to their own module
