+++ crates/git_hosting_providers/Cargo.toml
@@ -21,7 +21,6 @@ http_client.workspace = true
 regex.workspace = true
 schemars.workspace = true
 serde.workspace = true
-serde_derive.workspace = true
 serde_json.workspace = true
 settings.workspace = true
 url.workspace = true
